Extract ParamsValidator class
We were previously duplicating the behavior of validating params within
both, the template and document classes. This extracts the
ParamsValidator to encapsulate this.
