+++ pandadoc-api.gemspec
@@ -29,4 +29,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rspec', '~> 3.0'
   spec.add_development_dependency 'rubocop', '~> 0.58'
   spec.add_development_dependency 'webmock', '~> 3.4'
+  spec.add_development_dependency 'pry'
 end
